add: electron layer re-structure
