+++ src/ipc/cfg.ipc.ts
@@ -1,0 +1,36 @@
+import { ipcMain } from "electron";
+import { readGameCfg, writeGameCfg, GameCfg } from "../services/cfg.service";
+import {
+  readAppTitleCfg,
+  updatePs1TitleCfg,
+} from "../services/apps.service";
+
+export function registerCfgIpc(): void {
+  ipcMain.handle(
+    "read-app-title-cfg",
+    async (_event, oplRoot: string, folder: string) => {
+      return readAppTitleCfg(oplRoot, folder);
+    }
+  );
+
+  ipcMain.handle(
+    "read-game-cfg",
+    async (_event, oplRoot: string, gameId: string) => {
+      return readGameCfg(oplRoot, gameId);
+    }
+  );
+
+  ipcMain.handle(
+    "write-game-cfg",
+    async (_event, oplRoot: string, gameId: string, entries: GameCfg) => {
+      return writeGameCfg(oplRoot, gameId, entries);
+    }
+  );
+
+  ipcMain.handle(
+    "update-ps1-title-cfg",
+    async (_event, launcherPath: string, newTitle: string, gameId?: string) => {
+      return updatePs1TitleCfg(launcherPath, newTitle, gameId);
+    }
+  );
+}
