+++ src/ipc/index.ts
@@ -1,0 +1,27 @@
+import { registerWindowIpc } from "./window.ipc";
+import { registerLibraryIpc } from "./library.ipc";
+import { registerDialogIpc } from "./dialog.ipc";
+import { registerArtworkIpc } from "./artwork.ipc";
+import { registerImportIpc } from "./import.ipc";
+import { registerRenameIpc } from "./rename.ipc";
+import { registerDeleteIpc } from "./delete.ipc";
+import { registerCfgIpc } from "./cfg.ipc";
+import { registerAppsIpc } from "./apps.ipc";
+import { registerVmcIpc } from "./vmc.ipc";
+import { registerZsoIpc } from "./zso.ipc";
+import { registerSettingsIpc } from "./settings.ipc";
+
+export function registerAllIpc(): void {
+  registerWindowIpc();
+  registerLibraryIpc();
+  registerDialogIpc();
+  registerArtworkIpc();
+  registerImportIpc();
+  registerRenameIpc();
+  registerDeleteIpc();
+  registerCfgIpc();
+  registerAppsIpc();
+  registerVmcIpc();
+  registerZsoIpc();
+  registerSettingsIpc();
+}
