+++ src/services/artwork.service.ts
@@ -1,0 +1,86 @@
+import * as fs from "fs/promises";
+import path from "path";
+import https from "https";
+import { createLogger, formatBytes } from "../logger";
+
+const log = createLogger("artwork");
+
+export async function downloadArtByGameId(
+  dirPath: string,
+  gameId: string,
+  system: "PS1" | "PS2" = "PS2",
+  saveAsName?: string,
+  artTypes?: string[]
+) {
+  const baseUrl = `https://raw.githubusercontent.com/Luden02/psx-ps2-opl-art-database/refs/heads/main/${system}`;
+  const types = artTypes ?? ["COV", "ICO", "SCR"];
+  const results: any[] = [];
+  const localName = saveAsName || gameId;
+
+  log.info(
+    `Downloading ${system} artwork for ${gameId} (${types.join(", ")}) into ${dirPath}`
+  );
+
+  for (const type of types) {
+    const fileName = `${gameId}_${type}.png`;
+    const url = `${baseUrl}/${gameId}/${fileName}`;
+    log.verbose(`GET ${url}`);
+
+    try {
+      const buffer = await new Promise<Buffer>((resolve, reject) => {
+        https
+          .get(url, (res) => {
+            if (res.statusCode !== 200) {
+              return reject(
+                new Error(`Failed to download ${fileName}: ${res.statusCode}`)
+              );
+            }
+            const data: Buffer[] = [];
+            res.on("data", (chunk) => data.push(chunk));
+            res.on("end", () => resolve(Buffer.concat(data)));
+          })
+          .on("error", reject);
+      });
+
+      const savePath = path.join(dirPath, `${localName}_${type}.png`);
+      await fs.writeFile(savePath, buffer);
+      log.verbose(`Saved ${type} artwork (${formatBytes(buffer.length)}) → ${savePath}`);
+      results.push({
+        name: localName,
+        type,
+        url,
+        savedPath: savePath,
+      });
+    } catch (err: any) {
+      log.verbose(`${type} artwork unavailable for ${gameId}: ${err.message}`);
+      results.push({
+        name: localName,
+        type,
+        url,
+        error: err.message,
+      });
+    }
+  }
+
+  const saved = results.filter((r) => r.savedPath).length;
+  log.info(`Artwork for ${gameId}: ${saved}/${types.length} file(s) downloaded`);
+  if (saved === 0) {
+    const msg = `No artwork found for ${gameId} in ${system} database.`;
+    log.warn(msg);
+    return { success: false, data: results, message: msg };
+  }
+  return { success: true, data: results };
+}
+
+export async function checkArtFilesExist(artDir: string, filenames: string[]) {
+  const existing: string[] = [];
+  for (const name of filenames) {
+    try {
+      await fs.access(path.join(artDir, name));
+      existing.push(name);
+    } catch {
+      // File does not exist — skip.
+    }
+  }
+  return existing;
+}
