+++ src/services/settings.service.ts
@@ -1,7 +1,7 @@
 import { app } from "electron";
 import fs from "fs";
 import path from "path";
-import { createLogger } from "./logger";
+import { createLogger } from "../logger";
 
 const log = createLogger("settings");
