+++ src/utils/cue-parser.ts
@@ -1,6 +1,6 @@
 import * as fs from "fs/promises";
 import path from "path";
-import { createLogger } from "./logger";
+import { createLogger } from "../logger";
 
 const log = createLogger("cue-parser");
