+++ src/utils/cue2pops.ts
@@ -8,7 +8,7 @@ import {
   CueSheet,
   CueTrack,
 } from "./cue-parser";
-import { createLogger, formatBytes } from "./logger";
+import { createLogger, formatBytes } from "../logger";
 
 const log = createLogger("cue2pops");
