+++ src/utils/fs-entry.ts
